Split live map queries off the Viewport snapshot
Viewport carries only frozen values now; the coordinate conversions and
rendered-feature queries move onto CameraState, where asking the live
map is what the receiver says it does.
