+++ demo-app/common/src/commonMain/kotlin/org/maplibre/compose/demoapp/benchmark/BenchmarkMap.kt
@@ -171,10 +171,9 @@ internal fun BenchmarkMap(state: DemoAppState, sheetInsets: WindowInsets = Windo
               session.pointerPx = Offset(change.position.x, change.position.y)
             }
             if (change.pressed && session.pin == null && scenario.usesGestures) {
-              val viewport = cameraState.viewport
-              if (viewport != null) {
+              if (cameraState.viewport != null) {
                 session.pin =
-                  viewport.positionFromScreenLocation(
+                  cameraState.positionFromScreenLocation(
                     with(density) { DpOffset(change.position.x.toDp(), change.position.y.toDp()) }
                   )
               }
@@ -236,7 +235,8 @@ internal fun BenchmarkMap(state: DemoAppState, sheetInsets: WindowInsets = Windo
 
 private fun samplePin(session: BenchmarkSession) {
   val pin = session.pin ?: return
-  val projected = session.cameraState.viewport?.screenLocationFromPosition(pin) ?: return
+  if (session.cameraState.viewport == null) return
+  val projected = session.cameraState.screenLocationFromPosition(pin)
   val px = with(session.density) { Offset(projected.x.toPx(), projected.y.toPx()) }
   session.gestures.onMapProjection(px.x.toDouble(), px.y.toDouble())
 }
@@ -244,7 +244,10 @@ private fun samplePin(session: BenchmarkSession) {
 private fun androidx.compose.ui.graphics.drawscope.DrawScope.drawTrail(session: BenchmarkSession) {
   val composePoint = session.pointerPx
   val pin = session.pin
-  val projected = pin?.let { session.cameraState.viewport?.screenLocationFromPosition(it) }
+  val projected =
+    pin
+      ?.takeIf { session.cameraState.viewport != null }
+      ?.let { session.cameraState.screenLocationFromPosition(it) }
   val mapPoint = projected?.let { with(session.density) { Offset(it.x.toPx(), it.y.toPx()) } }
   if (composePoint != null) {
     val arm = 12.dp.toPx()
