+++ lib/maplibre-compose/src/jsMain/kotlin/org/maplibre/compose/map/GlJsMapSession.kt
@@ -579,7 +579,6 @@ internal class GlJsMapSession(
         visibleBoundingBox = getVisibleBoundingBox(),
         visibleRegion = getVisibleRegion(),
         metersPerDpAtTarget = metersPerDpAtLatitude(camera.zoom, camera.target.latitude),
-        map = this,
       )
     }
